requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.1.0
binaryornot==0.4.4
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cryptography==38.0.1
csrf==0.1b1
dateparser==1.1.2
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
flask-buzz==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
flask-praetorian==1.4.0
Flask-WTF==1.0.1
frozenlist==1.3.1
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mongoengine==0.24.2
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
passlib==1.7.4
pendulum==2.1.2
py-buzz==2.1.3
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
python-binance==1.0.16
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
regex==2022.3.2
requests==2.28.1
six==1.16.0
text-unidecode==1.3
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.1
zipp==3.10.0
